Move common test setup to CandidateTestSetup
